backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
pydantic>=2.4.0
httpx>=0.25.0
pytest>=7.4.0
python-multipart>=0.0.6
python-dotenv>=1.0.0